NukeViet 4.0 Final

+++ modules/news/admin/content.php
@@ -177,7 +177,7 @@ if ($rowcontent['id'] > 0) {
     $page_title = $lang_module['content_edit'];
     $rowcontent['topictext'] = '';
 
-    $body_contents = $db->query('SELECT * FROM ' . NV_PREFIXLANG . '_' . $module_data . '_bodyhtml_' . ceil($rowcontent['id'] / 2000) . ' where id=' . $rowcontent['id'])->fetch();
+    $body_contents = $db->query('SELECT * FROM ' . NV_PREFIXLANG . '_' . $module_data . '_detail where id=' . $rowcontent['id'])->fetch();
     $rowcontent = array_merge($rowcontent, $body_contents);
     unset($body_contents);
 
@@ -250,24 +250,25 @@ foreach ($global_array_cat as $catid_i => $array_value) {
 
 if ($nv_Request->get_int('save', 'post') == 1) {
     $catids = array_unique($nv_Request->get_typed_array('catids', 'post', 'int', array()));
-    $id_block_content_post = array_unique($nv_Request->get_typed_array('bids', 'post', 'int', array()));
-
+    $rowcontent['listcatid'] = implode(',', $catids);
     $rowcontent['catid'] = $nv_Request->get_int('catid', 'post', 0);
 
-    $rowcontent['listcatid'] = implode(',', $catids);
-
+    $id_block_content_post = array_unique($nv_Request->get_typed_array('bids', 'post', 'int', array()));
     if ($nv_Request->isset_request('status1', 'post')) {
         $rowcontent['status'] = 1;
-    } //dang tin
+        //Dang tin
+    }
     elseif ($nv_Request->isset_request('status0', 'post')) {
         $rowcontent['status'] = 0;
-    } //cho tong bien tap duyet
+    }
     elseif ($nv_Request->isset_request('status4', 'post')) {
         $rowcontent['status'] = 4;
-    } //luu tam
+        //Luu tam
+    }
     else {
         $rowcontent['status'] = 6;
-    } //gui, cho bien tap
+        //Gui, cho bien tap
+    }
 
     $message_error_show = $lang_module['permissions_pub_error'];
     if ($rowcontent['status'] == 1) {
@@ -402,18 +403,20 @@ if ($nv_Request->get_int('save', 'post') == 1) {
         $rowcontent['keywords'] = implode(',', $keywords_return);
     }
 
-    if (empty($rowcontent['title'])) {
-        $error[] = $lang_module['error_title'];
-    } elseif (empty($rowcontent['listcatid'])) {
-        $error[] = $lang_module['error_cat'];
-    } elseif (trim(strip_tags($rowcontent['bodyhtml'])) == '' and ! preg_match("/\<img[^\>]*alt=\"([^\"]+)\"[^\>]*\>/is", $rowcontent['bodyhtml'])) {
-        $error[] = $lang_module['error_bodytext'];
-    }
+	if ($rowcontent['status'] != 4) {
+	    if (empty($rowcontent['title'])) {
+	        $error[] = $lang_module['error_title'];
+	    } elseif (empty($rowcontent['listcatid'])) {
+	        $error[] = $lang_module['error_cat'];
+	    } elseif (trim(strip_tags($rowcontent['bodyhtml'])) == '' and ! preg_match("/\<img[^\>]*alt=\"([^\"]+)\"[^\>]*\>/is", $rowcontent['bodyhtml'])) {
+	        $error[] = $lang_module['error_bodytext'];
+	    }
+	}
 
     if (empty($error)) {
-        $rowcontent['catid'] = in_array($rowcontent['catid'], $catids) ? $rowcontent['catid'] : $catids[0];
-        $rowcontent['bodytext'] = nv_news_get_bodytext($rowcontent['bodyhtml']);
-
+    	if(!empty($catids)) {
+    		$rowcontent['catid'] = in_array($rowcontent['catid'], $catids) ? $rowcontent['catid'] : $catids[0];
+    	}
         if (! empty($rowcontent['topictext']) and empty($rowcontent['topicid'])) {
             $weightopic = $db->query('SELECT max(weight) FROM ' . NV_PREFIXLANG . '_' . $module_data . '_topics')->fetchColumn();
             $weightopic = intval($weightopic) + 1;
@@ -470,7 +473,10 @@ if ($nv_Request->get_int('save', 'post') == 1) {
 
         // Xu ly anh minh hoa
         $rowcontent['homeimgthumb'] = 0;
-        if (! nv_is_url($rowcontent['homeimgfile']) and nv_is_file($rowcontent['homeimgfile'], $uploads_dir_user) === true) {
+		if(empty($rowcontent['homeimgfile']) AND ( $rowcontent['imgposition'] == 1 OR $rowcontent['imgposition'] == 2 ) ){
+			$rowcontent['homeimgfile'] = nv_get_firstimage($rowcontent['bodyhtml']);
+		}
+        if (! nv_is_url($rowcontent['homeimgfile']) and nv_is_file($rowcontent['homeimgfile'], NV_UPLOADS_DIR . '/' . $module_upload) === true) {
             $lu = strlen(NV_BASE_SITEURL . NV_UPLOADS_DIR . '/' . $module_upload . '/');
             $rowcontent['homeimgfile'] = substr($rowcontent['homeimgfile'], $lu);
             if (file_exists(NV_ROOTDIR . '/' . NV_FILES_DIR . '/' . $module_upload . '/' . $rowcontent['homeimgfile'])) {
@@ -535,10 +541,7 @@ if ($nv_Request->get_int('save', 'post') == 1) {
                 nv_insert_logs(NV_LANG_DATA, $module_name, $lang_module['content_add'], $rowcontent['title'], $admin_info['userid']);
                 $ct_query = array();
 
-                $tbhtml = NV_PREFIXLANG . '_' . $module_data . '_bodyhtml_' . ceil($rowcontent['id'] / 2000);
-                $db->query("CREATE TABLE IF NOT EXISTS " . $tbhtml . " (id int(11) unsigned NOT NULL, bodyhtml longtext NOT NULL, sourcetext varchar(255) NOT NULL default '', imgposition tinyint(1) NOT NULL default '1', copyright tinyint(1) NOT NULL default '0', allowed_send tinyint(1) NOT NULL default '0', allowed_print tinyint(1) NOT NULL default '0', allowed_save tinyint(1) NOT NULL default '0', gid mediumint(9) NOT NULL DEFAULT '0', PRIMARY KEY (id)) ENGINE=MyISAM");
-
-                $stmt = $db->prepare('INSERT INTO ' . $tbhtml . ' VALUES
+                $stmt = $db->prepare('INSERT INTO ' . NV_PREFIXLANG . '_' . $module_data . '_detail VALUES
 					(' . $rowcontent['id'] . ',
 					 :bodyhtml,
 					 :sourcetext,
@@ -556,10 +559,6 @@ if ($nv_Request->get_int('save', 'post') == 1) {
                 foreach ($catids as $catid) {
                     $ct_query[] = ( int )$db->exec('INSERT INTO ' . NV_PREFIXLANG . '_' . $module_data . '_' . $catid . ' SELECT * FROM ' . NV_PREFIXLANG . '_' . $module_data . '_rows WHERE id=' . $rowcontent['id']);
                 }
-
-                $stmt = $db->prepare('INSERT INTO ' . NV_PREFIXLANG . '_' . $module_data . '_bodytext VALUES (' . $rowcontent['id'] . ', :bodytext )');
-                $stmt->bindParam(':bodytext', $rowcontent['bodytext'], PDO::PARAM_STR, strlen($rowcontent['bodytext']));
-                $ct_query[] = ( int )$stmt->execute();
 
                 if (array_sum($ct_query) != sizeof($ct_query)) {
                     $error[] = $lang_module['errorsave'];
@@ -616,7 +615,7 @@ if ($nv_Request->get_int('save', 'post') == 1) {
                 nv_insert_logs(NV_LANG_DATA, $module_name, $lang_module['content_edit'], $rowcontent['title'], $admin_info['userid']);
 
                 $ct_query = array();
-                $sth = $db->prepare('UPDATE ' . NV_PREFIXLANG . '_' . $module_data . '_bodyhtml_' . ceil($rowcontent['id'] / 2000) . ' SET
+                $sth = $db->prepare('UPDATE ' . NV_PREFIXLANG . '_' . $module_data . '_detail SET
 					bodyhtml=:bodyhtml,
 					sourcetext=:sourcetext,
 					imgposition=' . intval($rowcontent['imgposition']) . ',
@@ -632,21 +631,23 @@ if ($nv_Request->get_int('save', 'post') == 1) {
 
                 $ct_query[] = ( int )$sth->execute();
 
-                $array_cat_old = explode(',', $rowcontent_old['listcatid']);
-                $array_cat_new = explode(',', $rowcontent['listcatid']);
+				if ($rowcontent_old['listcatid'] != $rowcontent['listcatid']) {
+                    $array_cat_old = explode(',', $rowcontent_old['listcatid']);
+                    $array_cat_new = explode(',', $rowcontent['listcatid']);
+				    $array_cat_diff = array_diff($array_cat_old, $array_cat_new);
+	                foreach ($array_cat_diff as $catid) {
+	                	if (!empty($catid)) {
+	                		$ct_query[] = $db->exec('DELETE FROM ' . NV_PREFIXLANG . '_' . $module_data . '_' . $catid . ' WHERE id = ' . intval($rowcontent['id']));
+	                	}
+	                }
+				}
 
-                $array_cat_diff = array_diff($array_cat_old, $array_cat_new);
-                foreach ($array_cat_diff as $catid) {
-                    $ct_query[] = $db->exec('DELETE FROM ' . NV_PREFIXLANG . '_' . $module_data . '_' . $catid . ' WHERE id = ' . $rowcontent['id']);
-                }
-                foreach ($array_cat_new as $catid) {
-                    $db->exec('DELETE FROM ' . NV_PREFIXLANG . '_' . $module_data . '_' . $catid . ' WHERE id = ' . $rowcontent['id']);
-                    $ct_query[] = $db->exec('INSERT INTO ' . NV_PREFIXLANG . '_' . $module_data . '_' . $catid . ' SELECT * FROM ' . NV_PREFIXLANG . '_' . $module_data . '_rows WHERE id=' . $rowcontent['id']);
-                }
-
-                $sth = $db->prepare('UPDATE ' . NV_PREFIXLANG . '_' . $module_data . '_bodytext SET bodytext=:bodytext WHERE id =' . $rowcontent['id']);
-                $sth->bindParam(':bodytext', $rowcontent['bodytext'], PDO::PARAM_STR, strlen($rowcontent['bodytext']));
-                $ct_query[] = ( int )$sth->execute();
+				foreach ($catids as $catid) {
+				    if (!empty($catid)) {
+				        $db->exec('DELETE FROM ' . NV_PREFIXLANG . '_' . $module_data . '_' . $catid . ' WHERE id = ' . $rowcontent['id']);
+				        $ct_query[] = $db->exec('INSERT INTO ' . NV_PREFIXLANG . '_' . $module_data . '_' . $catid . ' SELECT * FROM ' . NV_PREFIXLANG . '_' . $module_data . '_rows WHERE id=' . $rowcontent['id']);
+				    }
+				}
 
                 if (array_sum($ct_query) != sizeof($ct_query)) {
                     $error[] = $lang_module['errorsave'];
@@ -684,6 +685,7 @@ if ($nv_Request->get_int('save', 'post') == 1) {
                 $keywords = array_unique($keywords);
 
                 foreach ($keywords as $keyword) {
+                    $keyword = str_replace('&', ' ', $keyword);
                     if (! in_array($keyword, $array_keywords_old)) {
                         $alias_i = ($module_config[$module_name]['tags_alias']) ? change_alias($keyword) : str_replace(' ', '-', $keyword);
                         $alias_i = nv_strtolower($alias_i);
@@ -746,20 +748,21 @@ if ($nv_Request->get_int('save', 'post') == 1) {
                 $url = NV_BASE_ADMINURL . 'index.php?' . NV_LANG_VARIABLE . '=' . NV_LANG_DATA . '&' . NV_NAME_VARIABLE . '=' . $module_name;
                 $msg1 = $lang_module['content_saveok'];
                 $msg2 = $lang_module['content_main'] . ' ' . $module_info['custom_title'];
-                redriect($msg1, $msg2, $url, $module_data . '_bodyhtml');
+                redriect($msg1, $msg2, $url, $module_data . '_detail');
             }
         }
     } else {
         $url = 'javascript: history.go(-1)';
         $msg1 = implode('<br />', $error);
         $msg2 = $lang_module['content_back'];
-        redriect($msg1, $msg2, $url, $module_data . '_bodyhtml', 'back');
+        redriect($msg1, $msg2, $url, $module_data . '_detail', 'back');
     }
     $id_block_content = $id_block_content_post;
 }
 
 $rowcontent['hometext'] = nv_htmlspecialchars(nv_br2nl($rowcontent['hometext']));
 $rowcontent['bodyhtml'] = htmlspecialchars(nv_editor_br2nl($rowcontent['bodyhtml']));
+$rowcontent['alias'] = ($rowcontent['status'] == 4 and empty($rowcontent['title'])) ? '' : $rowcontent['alias'];
 
 if (! empty($rowcontent['homeimgfile']) and file_exists(NV_UPLOADS_REAL_DIR . '/' . $module_upload . '/' . $rowcontent['homeimgfile'])) {
     $rowcontent['homeimgfile'] = NV_BASE_SITEURL . NV_UPLOADS_DIR . '/' . $module_upload . '/' . $rowcontent['homeimgfile'];
@@ -1025,7 +1028,6 @@ $contents .= $xtpl->text('main');
 if ($rowcontent['id'] > 0) {
     $op = '';
 }
-
 include NV_ROOTDIR . '/includes/header.php';
 echo nv_admin_theme($contents);
 include NV_ROOTDIR . '/includes/footer.php';
